feat: upgrade to Spring Boot 3.0 / Kotlin 1.8

+++ search-service/src/main/kotlin/com/egm/stellio/search/util/DBQueryUtils.kt
@@ -45,7 +45,7 @@ suspend fun DatabaseClient.GenericExecuteSpec.execute(): Either<APIException, Un
         .awaitFirst()
 
 suspend fun DatabaseClient.GenericExecuteSpec.executeExpected(
-    f: (value: Int) -> Either<APIException, Unit>
+    f: (value: Long) -> Either<APIException, Unit>
 ): Either<APIException, Unit> =
     this.fetch()
         .rowsUpdated()
